Not Moving Switch
Added a not moving toggle in the extra space (we did have a lot of extra) and made it functional. Also made the status text not use a hardcoded string and take up less space. I couldn't figure out/don't think you can make the toggle part on the bottom without completely making it customizing /redoing it.