Last of the refactoring - coding standards
Mostly refactored all of the onCreates to call a number of functions to do all of the initializing.  This makes onCreate WAY easier to read and encapsulates each component functionality into its own function.  This looks messy but I just MOVED code from onCreate to the function.

Changed the way we call (and return) from Settings page.  This is WAY nicer and the way Android is meant to do it.

Added the skeleton for Achievements to pop up.  It's all ignored right now.

Also fixed a few bugs (see list)
teamToScout spinner - the loadTeamData() was mostly a re-write to handle all situations.  We had this code in 3-4 different spots in the activity.  really hard to debug.