fix #336
First, we weren't handling "Not Moving" at all properly like the other toggles.  Added the variables for that.

Secondly, we didn't reset the current menu group's "last event" correctly if it was the last (first?) one.  So if you undo Climb->Start, we removed if from the logging, but since we didn't find a "previous" climbing event, we just left it along (as climb Start, which isn't correct).

Thirdly, checking if we have NO previous event was wrong since the 2nd part of the OR was always true.  oops.

Cleaned up starting position order (per Erica)
Cleaned up EventIDs for toggle switches (per new ranges I made for ReefScape)