fixes
fix #348
fix #346

also fixed a crash loading QR codes that are for a different device.  shouldn't happen, but can happen (especially when debugging and pushing files onto the device).  now we only show matches that also match the device id in SubmitData so that QRCodes don't crash (they used the device id in the filename to load)