Fixed the undo issue
This is a bit involved.  First, I was able to remove all of the special coding for tracking the toggles (sequence number - so we knew how to link them via PrevSeq number).  For that, I created an EventGroup per toggle.  These are not FOP (field of play) so they won't show up on a menu.  We track previous sequence per EventGroup, so this means we get the toggles "for free" now.

I moved all of the tracking of "CurrentEvent" and "PreviousSeq" to the logger (from Match).  Since we can start with a game piece, we needed this before the match activity was created and the ArraysFill in Match, messed it up.  Now it's in Logger where it belongs.  The rest of the app shouldn't have to worry about keeping it accurate.

Removed the global "is_start_of_seq" since it also now depends on which group it is (we can have ore than one start).  Made an Event function to return whether an event is the start of a seq (we had the data anyways) and use that instead.

Logger: added a few "Objects.requireNonNulls" to remove warnings.

Logger: LookupEvent was never used, so removed it.

Logger: biggest changes are in LogEvent and Undo.  It's quite a lot cleaner to read, and debug.  It not cleans up the current_event and previous_seq to get the logging and menus correct.  Then it figures out what to show in the status field as the "last event" and gives that back to the caller (so we can format the status properly).