remaining achievements
added a field in Competitions.csv to track if we attended the competition or not.
updated how we load competitions as well as the competition class

in postMatch, simplified Adriel's code to be generic (needed this for dcmp anyways).

in constants, removed newton (it's not an achievement we have) and changed einstein to use an arraylist so we can use the same rule as worlds (reuse!)

the rest should be clear.