fixed it!
Cleaned up the calls to Logger:
no longer need to pass in if this starts a sequence.  Handled now inside the logger.
no longer need to pass in x=0, y=0 for toggles.  Handled now inside the logger.
Exception is StartingNote since it need to override the "time" parameter.

FIX:
When we end the match with a toggle on, we LogEvent to turn it off, then we uncheck the toggle (which then calls LogEvent to log turning it off) resulting in a double call.  removed the extra call to LogEvent!

FIX: when undo'ing an event, we also need to update the internal structure tracking the previous_seq for any group if it's above where we deleted from (lastIndex).  Otherwise the values here won't be accurate and we can use them to link the next one.