Fixed bugs
Fixed incrementing the match number.  Now it ONLY does it when you hit "next match".  Also (thanks Adriel) if you jump to "resubmit data" we pre-decrement the next match number knowing that we'll increment it going back to the pre-match activity.  Tested scenarios and it works well now.

We load the matches a little differently now.  We DO rely on INDEX # = match # and assumed it was sequential-ish.  Now we don't and build the ArrayList properly - still ending up with Index = match #.  But not the ordering doesn't matter.  Tested and looks good.

Also cleaned up (a bit) how we handle "no match" data.  No CONSTANT needed, just a named fn() call.  (reused the "csv" one since it was only called for a no-match anyways).

Fixed a small bug found in testing where if you have a valid match, with a team and the team name pops up, then you move to an invalid match, we used to blank out the list of teams, but the old team name remained (WRONG!).  We not blank it out.