fixed bugs!
Events.csv: added a Park option to the context menu.  Still need Tableau changes.

Match.java: sorry for the renames.  StartTime and the Transmit* ones were not Camel-Case like the other Globals.

Added check for FieldOrientation.  If it's Automatic, do what we did before, but if it's set to Red/Blue, just set the currentOrientation and Image to the right values.

QRCode.java: Made a few things Static so hopefully they don't build on each other every call (not sure it helped).  Also using a StringBuilder (recommended by Android Studio).  Clearing out the qrFileString so the GC can clean it up.

Settings: added an option to set the field orientation to "Auto", "Blue" or "Red".

Globals: removed deviceAddress - never used (was for BlueTooth)