show only match types that we have files for
fixes #434

SubmitData: instead of using the "Data" list of all match types, use the file_list we generate based off of what's actually on disk for this competition to get a list of possible match types that can actually be picked from.  Scan through and store the shortform into a hashmap (just easier than having to check if we found it already), convert to an arraylist to use in the spinner.  Also adjust how we choose a default spinner item based on the match type we (might have) just scouted.

MatchTypes: needed a new function to convert ShortForm to Description (for the spinner).