Converting TEAM number to String
Changed all references to Team number to be a string and ran through some full tests to validate.

In Matches.java also no longer assume if the alliance isn't blue, it must be red.  I explicitly check for Red, otherwise return nothing.  This is safer to do in case there's no match schedule - not all 6 teams are RED!

Lastly, in Logger.java, since we're now writing to the same file, i noticed that opening the FileOutputStream (fos) twice, overwrites the data that was written.  Oops.  So now we open the filestream once, and use it twice (it's slightly more efficient to do it this way anyways.  Tested it and the file now has both the "D" and "E" data.