Handle a trimmed Competition list file
Changed the achievement data to be a HashMap rather than an index to avoid the OOB exception.  Changed all references to use the HashMap rather than the array.

Also needed to change the Settings page to handle the default/current competition no longer being in the competition list (say it was trimmed out) - before we set the spinner selection, just make sure it's a valid one.