Pop Achievement better
We no longer have one achievement pop the next one (the timing of this and keeping the variables in sync while using the debugger caused a timing issue and results in an OOB exception if you're in the debugger while trying to pop >1 achievement).

Instead, we schedule ALL achievements at the get go, with a delayed start to stagger them out.  Just needed to create a constructor that took in the achievement index that it needed to worry about.

Also in Logger, added a check that output_df isn't null before trying to run listFiles() - it shouldn't be, but just in case it is, this will result in a null pointer exception.  So best to protect against it.

fixes #482