remove ID from MatchTypes
Since we used an ArrayList and the ID as the index, we lost that ability.   So i converted this to a HashMap and then Matches became a HashMap of HashMaps.  So a lot of code changed to do the same thing but with a HashMap.  It meant we didn't need some functions convering MatchTypeId's and we didn't need a "NoMatch" record since the HashMap doesn't care about that.
Also added in a bunch of "Null" checks.

In Submitdata: also moved pop_list to a local variable - should have done this in a previous merge when fixing how we pop achievements.  Not an error, just a compiler warning and cleaner code.

Also changed the App versioning in build.gradle.kts to drop the "Build Version" and allow for 2-digits for Minor and Patch values.

fixes #485