Fix incorrect logging of files
Logger: split out "Writing of files" from "Closing the logger".  So we can "close" the logger without writing out files.

in addition, save off filename to be used when the logger is instantiated (in constructor) rather than at the end when writing out files.  Although this bug isn't an issue with the other fixes, it's still better to determine the filename upfront.

We don't need the "clear()" logger function now, as this is what close() does.

The "assert" line is just to get rid of a compiler warning.

PreMatch - if we're going to the SubmitData page because of a reset, CLOSE the logger first.  (then in submitdata we won't write anything out because the logger doesn't exist.

fixes #457