Fix
Fixes #488
SubmitData: Only pop achievement if there's a valid logger.
SubmitData: found bug that we need to use the TransmitMatchType directly since it's now already the shortform, and not an integer index.

Logger: found bug that we need to create the filename using the Globals.MatchType directly since it's now already the shortform, and not an integer index.