Clicking too fast crashes
fixes #456
Match: add a check for a "no event" and return gracefully.  If you click too fast, we can call this with the default "no event id" (=999) and getting the group ID for that is -1, which will cause an OOB later in the function.  We'll get here if we can't match the text of the event to an ID (which can happen if you click too fast, and we get a timing issue where what you clicked on ("L4"?) is no longer in the list of available events ('cause you click on another menu too fast).

Settings: also noticed this line was still expecting an INT for scouting team when it's saved as a string.  Going back into Settings crashed.  Missed change from earlier check-in.

Updated gradle to 8.13.1