fix a startup crash
Happens only first time you run it, but we try to load the data twice (i'm thinking it's a callback from checking storage permissions.

EventGroup wasn't static, but MaxEventGroups was set to 6 the first time we loaded data.  then on the 2nd attempt, the datastructure (not being static) was a new instance and was zero size.  since MaxEventGroup was big enough, we didn't add a row, and then tried to access it and got an OOB.  I also changed the "adding a new row" to be based on the actual size to avoid this in the future.