fixes
Matches: removed an unused constructor and then made the member variables final and removed initialization (cleaned up code)

Changed CPR_Chronometer to use SystemClock.elapsedRealtime() and that seemed to be needed for it to work in the widget.  Made chronometer final as well so you can't change which chronometer is used for the CPR version.  :)

Match.xml - removed the TextView we used to show the timer, and now just use the Chronometer object (one less widget to update and deal wtih)