Added a transition event for MatchPhase
fixes #329

Events.csv: added a column for which event to use in case the MatchPhase transitions to something else.  (going from AUTO to TELEOP)

AppLaunch: handle the extra column we read in from Events.csv

Match.java: removed the "static" - compiler warning - possible memory leak

Constants.java: added a constant to flag there the transition even does not exist (DNE)

Events.java: added the extra field (transition_event) and in "getNextEvents" (which is called when building the dropdown menus) see if the MatchPhase has changed AND there's a transition event, and if so, change to using the new eventId.