improve context menus
fixes #505

match.xml:
removed view and relative layout that aren't needed anymore (the relative layout wasn't needed anyways)

Events.java
we now store the list of next-event-ids as well as the descriptions since we populate the context menus using the ID's (under the covers) as well as the text descriptions.
added functions to get the list and deleted old functions no longer used.

Match.java
we now build the whole menu/submenu structure each time you click on the screen to log an event - utilizing the next-event-ids for each group.
This also means we don't have to build our own submenu system and track where we are - we use it all built-in.
We also no longer need to look up the eventID based on the text used in the menu - we can store the EventID inside the menu structure for easy retrieval.
Also moved the check for "isPractice" INSIDE the setEventStatus fn to make it cleaner looking (and so if we use it anywhere in the code we don't forget to make the check).