Misc Fixes
AppLaunch: started to experiment with some animation.  The "Scouting App" text now fades in.

Match.java: I can't use automation on a button background (was hoping to use this to avoid using a timer) but we can do a "soft" color transition to fade in the red color.  I could fade it out, but this pattern works well.  (a fade_out would make it "pulse" and could work, but I like this)

Constants.java - removed some unused values

*.xml (except fade_in.xml) - removed layout elements used to create a gap inbetween rows.  Found you can use marginTop to do the same thing.  Less elements is cleaner.